requirements.txt
geopy==2.4.1
requests==2.32.3
google-genai==1.15.0
pandas==2.2.3
poetry==1.8.4
openai==1.90.0
tiktoken==0.9.0
anthropic==0.54.0
torch==2.7.1
PyYAML==6.0.2
litellm==1.73.0
pydantic-ai==0.3.2